% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hargreaves}
\alias{hargreaves}
\title{Hargreaves equation for evapotranspiration}
\usage{
hargreaves(S0, tmean, tmin, tmax)
}
\arguments{
\item{S0}{numeric}

\item{tmean}{degrees C}

\item{tmin}{degrees C}

\item{tmax}{degrees C}
}
\value{
evapotranspiration (ETP) in mm/day
}
\description{
Per Hargreaves & Samani 1985. (Shuttleworth 1993 seems to incorrectly omit the exponent)
}
