% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{latitude}
\alias{latitude}
\title{Create a latitude raster}
\usage{
latitude(template, already_latlong = FALSE)
}
\arguments{
\item{template}{a raster layer}

\item{already_latlong}{leave as F unless rasters are already in lat-long projection}
}
\value{
A raster layer with values representing degrees latitude
}
\description{
Create a latitude raster
}
