% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{monthly_S0}
\alias{monthly_S0}
\title{Mean S0 for a month of the year}
\usage{
monthly_S0(month, latitude)
}
\arguments{
\item{month}{integer from 1 to 12}

\item{latitude}{decimal degrees}
}
\value{
mean S0 for a month of the year
}
\description{
Mean S0 for a month of the year
}
